Remove unnecessary ModalAuthor
