--- library-valcon-internship/src/components/Modals/ModalAddAuthor/ModalAddAuthor.css
@@ -1,7 +1,0 @@
-.add-author-close-modal {
-  background-color: red;
-  color: white;
-  align-self: flex-end;
-  border-radius: 15px;
-  padding: 0.5rem;
-}
