+++ library-valcon-internship/src/components/Search/Search.tsx
@@ -7,7 +7,7 @@ import { VscFilter } from 'react-icons/vsc'
 import Filter from '../../models/Filter'
 import Sort from '../../models/Sort'
 import Where from '../../models/Where'
-import Modal from '../Modals/Modal'
+import Modal from '../Modals/Modal/Modal'
 import ModalFilter from '../Modals/ModalFilter/ModalFilter'
 import ModalSort from '../Modals/ModalSort/ModalSort'
 import './Search.css'
